disease information as nested object list
This is based on the original ga4gh metadata task team JS object representation of fictional disease/state annotations. Here, the idea is to use an object for each identified disease, which then collects annotations in the form of ontology end points, classifications or clinical parameters (but not complete representations of medical records etc.).